Move MessageAndUserFromWebhookJSON and PostsFromVKCreator to separate services folder
