+++ tgbot/services/MessageAndUserFromWebhookJSON.py
@@ -1,4 +1,10 @@
-from .models import User, Message
+from tgbot.models import User, Message, Group, Post
+from datetime import datetime
+
+from utils.getCategory import get_food_category, all_cats
+from utils.getCity import get_city, get_metro_station
+from utils.getIsBooked import get_is_booked
+from utils.getIsLost import get_is_lost
 from datetime import datetime
 
 # Класс для сохранения информации о юзере и сообщении в БД
